refactor(extensions): modular
